refactor: re-add missing strategies, make them serializable
