+++ generator/accessor/src/main/kotlin/me/kcra/takenaka/generator/accessor/naming/StandardNamingStrategies.kt
@@ -18,11 +18,7 @@
 package me.kcra.takenaka.generator.accessor.naming
 
 import me.kcra.takenaka.generator.accessor.GeneratedClassType
-import me.kcra.takenaka.generator.accessor.GeneratedMemberType
 import me.kcra.takenaka.generator.accessor.model.ClassAccessor
-import me.kcra.takenaka.generator.accessor.model.ConstructorAccessor
-import me.kcra.takenaka.generator.accessor.model.FieldAccessor
-import me.kcra.takenaka.generator.accessor.model.MethodAccessor
 
 /**
  * Standard [NamingStrategy] implementations.
@@ -38,6 +34,20 @@ object StandardNamingStrategies {
      * but different package; use [resolveSimpleConflicts].**
      */
     val SIMPLE: NamingStrategy = SimpleNamingStrategy()
+
+    /**
+     * A fully-qualified class name strategy, based on [SIMPLE].
+     *
+     * Uses the entire mapped name, including the full package.
+     */
+    val FULLY_QUALIFIED: NamingStrategy = FullyQualifiedNamingStrategy()
+
+    /**
+     * A semi-qualified class name strategy, based on [SIMPLE].
+     *
+     * Uses the entire mapped name, including the package, but with `net.minecraft`/`com.mojang` removed.
+     */
+    val SEMI_QUALIFIED: NamingStrategy = SemiQualifiedNamingStrategy()
 }
 
 /**
@@ -66,64 +76,6 @@ fun NamingStrategy.resolveSimpleConflicts(): NamingStrategy {
     return SimpleConflictResolvingNamingStrategy(this)
 }
 
-private class SimpleNamingStrategy : NamingStrategy {
-    override fun klass(model: ClassAccessor, type: GeneratedClassType): String {
-        val simpleName = model.internalName.substringAfterLast('/')
-
-        return when (type) {
-            GeneratedClassType.MAPPING -> "${simpleName}Mapping"
-            GeneratedClassType.ACCESSOR -> "${simpleName}Accessor"
-            else -> throw UnsupportedOperationException("$type is not a contextual type")
-        }
-    }
-
-    override fun klass(type: GeneratedClassType): String {
-        return when (type) {
-            GeneratedClassType.LOOKUP -> "Mappings"
-            else -> throw UnsupportedOperationException("$type is not a context-free type")
-        }
-    }
-
-    override fun field(model: FieldAccessor, index: Int): String {
-        val fieldName = "FIELD_${model.upperName}"
-        if (index != 0) {
-            return "${fieldName}_$index"
-        }
-
-        return fieldName
-    }
-
-    override fun fieldHandle(model: FieldAccessor, index: Int, mutating: Boolean): String {
-        val fieldName = field(model, index)
-        if (mutating) {
-            return "${fieldName}_SETTER"
-        }
-
-        return "${fieldName}_GETTER"
-    }
-
-    override fun constant(model: FieldAccessor, index: Int): String {
-        return field(model, index)
-    }
-
-    override fun constructor(model: ConstructorAccessor, index: Int): String {
-       return "CONSTRUCTOR_$index"
-    }
-
-    override fun method(model: MethodAccessor, index: Int): String {
-        val methodName = "METHOD_${model.upperName}"
-        if (index != 0) {
-            return "${methodName}_$index"
-        }
-
-        return methodName
-    }
-
-    override fun member(type: GeneratedMemberType): String {
-        return type.name
-    }
-}
-
 private class PrefixingNamingStrategy(next: NamingStrategy, val basePackage: String) : ForwardingNamingStrategy(next) {
     override fun klass(model: ClassAccessor, type: GeneratedClassType): String {
         return "$basePackage.${super.klass(model, type)}"
@@ -132,11 +84,16 @@ private class PrefixingNamingStrategy(next: NamingStrategy, val basePackage: Str
     override fun klass(type: GeneratedClassType): String {
         return "$basePackage.${super.klass(type)}"
     }
+
+    companion object {
+        private const val serialVersionUID = 1L
+    }
 }
 
 private data class TypedName(val name: String, val type: GeneratedClassType)
 
 private class SimpleConflictResolvingNamingStrategy(next: NamingStrategy) : ForwardingNamingStrategy(next) {
+    @Transient
     private val simpleNames = mutableMapOf<TypedName, MutableMap<String, Int>>() // {simple name + type: {fq name: occurrence index}}
 
     override fun klass(model: ClassAccessor, type: GeneratedClassType): String {
@@ -151,5 +108,9 @@ private class SimpleConflictResolvingNamingStrategy(next: NamingStrategy) : Forw
             return super.klass(ClassAccessor(model.name + index, model.fields, model.constructors, model.methods, model.requiredTypes), type)
         }
         return super.klass(model, type)
+    }
+
+    companion object {
+        private const val serialVersionUID = 1L
     }
 }
